add footer and memberdetail
