+++ src/components/Main/member/MemberCard.jsx
@@ -1,10 +1,9 @@
 import * as React from 'react';
 import { useContext, useState } from 'react';
+import MemberCardDetail from './MemberCardDetail';
 import styled from '@emotion/styled';
 import PropTypes from 'prop-types';
-import { NavLink } from 'react-router-dom';
 import Card from '@mui/material/Card';
-import { useNavigate } from 'react-router';
 import CardMedia from '@mui/material/CardMedia';
 import Typography from '@mui/material/Typography';
 import ButtonBase from '@mui/material/ButtonBase';
@@ -14,9 +13,6 @@ import MemberDeleteBtn from './MemberDeleteBtn';
 import { MemberEditBtn } from './MemberEditBtn';
 import { useCookies } from 'react-cookie';
 import { UserContext } from '../../../context/UserContext';
-import { useUserPostDetailApi } from '../../../hooks/Member';
-import ImageList from '@mui/material/ImageList';
-import ImageListItem from '@mui/material/ImageListItem';
 
 const SkillBtn = styled.button`
     height: 22px;
@@ -47,24 +43,7 @@ export default function MemberCard({
     const { user } = useContext(UserContext);
     const [cookies] = useCookies(['id']);
     const [isVisible, setIsVisible] = useState(false);
-    const projectList = [];
-    let projectCount = 0;
-    const navigate = useNavigate();
-    const navigateToProject = (index) => {
-        console.log('index', index);
-        // navigate(`/maintab/projectdetail/{index}`);
-    };
-    if (projects) {
-        for (let i = 0; i < projects.length; i++) {
-            if (projects[i] === null) {
-                projectCount = i;
-                break;
-            }
-            const [detail] = useUserPostDetailApi(projects[i]);
-            projectList.push([detail.mainimage, detail._id, detail.tags, detail.title]);
-        }
-    }
-    console.log('tag', projectList);
+
     return (
         <>
             <ButtonBase
@@ -191,76 +170,7 @@ export default function MemberCard({
                     </Grid>
                 </Card>
             </ButtonBase>
-            {isVisible ? (
-                <Card
-                    style={{
-                        width: '350px',
-                        marginTop: -25,
-                        borderRadius: 20,
-                        paddingTop: '40px',
-                        paddingBottom: '20px',
-                        paddingLeft: '8px',
-                        paddingRight: '8px',
-                        backgroundColor: '#303030',
-                        zIndex: '-1'
-                    }}
-                >
-                    <Box
-                        sx={{
-                            width: '100%',
-                            marginTop: '10px',
-                            textAlign: 'left'
-                        }}
-                    >
-                        <Typography
-                            sx={{ fontWeight: '700', fontSize: '0.75rem' }}
-                            variant="caption"
-                        >
-                            {' '}
-                            진행한 프로젝트{' '}
-                        </Typography>
-                        <span style={{ fontSize: '0.563rem' }}> &nbsp; 총 {projectCount}개</span>
-                        <ImageList
-                            style={{
-                                display: 'grid',
-                                gridAutoFlow: 'column'
-                            }}
-                        >
-                            {projectList.map((item) => (
-                                <ImageListItem
-                                    key={item[1]}
-                                    style={{
-                                        borderRadius: '15px',
-                                        border: '1px solid red',
-                                        width: '125px',
-                                        height: '125px',
-                                        margin: '3px'
-                                    }}
-                                    // onClick={navigateToProject(item[1])}
-                                >
-                                    <NavLink to={'/maintab/projectdetail/' + item[1]}>
-                                        <img
-                                            src={item[0]}
-                                            alt={item[3]}
-                                            // onClick={console.log('click', item[1])}
-                                        />
-                                    </NavLink>
-                                </ImageListItem>
-                            ))}
-                        </ImageList>
-                        <Typography
-                            sx={{ fontWeight: '700', fontSize: '0.75rem' }}
-                            variant="caption"
-                        >
-                            {' '}
-                            Skill Set
-                        </Typography>
-                        {projectList.map((item) => (
-                            <SkillBtn key={item[1]}>{item[2]}</SkillBtn>
-                        ))}
-                    </Box>
-                </Card>
-            ) : null}
+            {isVisible ? <MemberCardDetail projects={projects} /> : null}
         </>
     );
 }
